arreglando migraciones

+++ database/migrations/2021_11_17_115803_create_anotaciones_table.php
@@ -17,7 +17,7 @@ class CreateAnotacionesTable extends Migration
             $table->id();
             $table->text('Descripcion_Anotacion');
             $table->string('Tipo_Anotacion');
-            $table->unsignedInteger('ID_Asignatura');
+            $table->unsignedBigInteger('ID_Asignatura');
             $table->foreign('ID_Asignatura')->references('id')->on('asignaturas')->onDelete('cascade')->onUpdate('cascade');
             $table->unsignedInteger('Rut_Alumno');
             $table->foreign('Rut_Alumno')->references('Rut_Alumno')->on('usuario_alumnos')->onDelete('cascade')->onUpdate('cascade');
